feat(vite-plugin-angular): merge vfile export with metadata export
