+++ apps/ng-app/src/app/models.ts
@@ -1,0 +1,6 @@
+export interface PostAttributes {
+  title: string;
+  slug: string;
+  description: string;
+  coverImage: string;
+}
