+++ apps/ng-app/src/app/pages/posts.[slug].page.analog
@@ -5,31 +5,25 @@ import {
 } from '@analogjs/content';
 import { MarkdownComponent } from '@analogjs/content' with { analog: 'imports' }
 import { RouteMeta } from '@analogjs/router';
-import { AsyncPipe, JsonPipe, NgFor, NgIf } from '@angular/common' with { analog: 'imports'};
-import { Component, inject } from '@angular/core';
+import { effect, inject } from '@angular/core';
+import { toSignal } from '@angular/core/rxjs-interop';
 import { map } from 'rxjs';
 
-import { PostAttributes } from './models';
+import { PostAttributes } from '../models';
 
 const renderer = inject(ContentRenderer);
 const post$ = injectContent<PostAttributes>();
 
-const toc$ = post$.pipe(
-  map(() => {
-    return renderer.getContentHeadings();
-  })
-);
+const postSignal = toSignal(post$);
+
+effect(() => {
+  console.log(postSignal());
+})
 </script>
 
-<template><ng-container *ngIf="post$ | async as post">
-  <h1>{{ post.attributes.title }}</h1>
-  <div *ngIf="toc$ | async as toc">
-    <ul>
-      <li *ngFor="let item of toc">
-        <a href="#{{ item.id }}">{{ item.text }}</a>
-      </li>
-    </ul>
-  </div>
-
-  <analog-markdown [content]="post.content"></analog-markdown>
-</ng-container></template>
+<template>
+  @if(postSignal(); as post){
+    <h1>{{ post.attributes.title }}</h1>
+    <analog-markdown [content]="post.content"></analog-markdown>
+  }
+</template>
